+++ src/app/_services/pdf.service.ts
@@ -1,5 +1,7 @@
 import { Injectable } from '@angular/core';
 
+import  * as moment from "moment";
+
 // import * as jsPDF from 'jspdf';
 declare let html2pdf;
 declare var jsPDF : any;
@@ -31,16 +33,37 @@ export class PdfService {
   }
 
   generatePdfByAutotable(title : string,subtitle : string){
+    let currentDate = moment(new Date()).format("MMMM Do, YYYY");
     let doc = new jsPDF();
         doc.setFontSize(20);
         doc.setTextColor(40);
-        doc.setFontStyle('normal');
+        doc.setFontStyle('Times New Roman');
         doc.text(title,14,16);
         doc.setFontSize(10);
+        doc.text(currentDate, 150, 16);
         doc.text(subtitle,14,22);
     let elem = document.getElementById('basic-table');
     let res = doc.autoTableHtmlToJson(elem);
-        doc.autoTable(res.columns,res.data,{startY: 25});
+        doc.autoTable(res.columns,res.data,{
+          theme:'grid',
+          startY: 25, 
+          showHeader:'everyPage',
+          styles: { font :'Times New Roman', fontSize: 10  },
+          drawRow: function(row,data) {
+            let tdClass = row.raw[0].className;
+            let tdText  = row.raw[0].innerText;
+            if(tdClass === 'category'){
+              row.cells[0].styles.fontSize = 12;
+              row.cells[0].styles.fontStyle = 'bold';
+            }
+            else if(tdClass === 'subCategory'){
+              row.cells[0].styles.cellPadding = 3;
+            }
+            
+          }
+        });
+
+        // console.log(res.data[0][0].className);
 
         let iframe = document.createElement('iframe');
         iframe.setAttribute('style','position:relative;right:0; top:0; bottom:0; height:100%; width:100%;');