+++ src/app/features/expenses-report/expenses-report.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { SalesIndividual, SalesByCategory } from './sales-report.model';
+import { ExpensesIndividual, ExpensesByCategory } from './expenses-report.model';
 
-import { SalesReportService } from './sales-report.service';
+import { ExpensesReportService } from './expenses-report.service';
 
 import { ProgressDialogService } from './../../_services/progress-dialog.service';
 import { ErrorHandlerService } from './../../_services/error-handler.service';
@@ -11,17 +11,17 @@ import { ErrorHandlerService } from './../../_services/error-handler.service';
 import * as moment from "moment";
 
 @Component({
-  selector: 'app-sales-report',
-  templateUrl: './sales-report.component.html',
-  styleUrls: ['./sales-report.component.scss']
+  selector: 'app-expenses-report',
+  templateUrl: './expenses-report.component.html',
+  styleUrls: ['./expenses-report.component.scss']
 })
-export class SalesReportComponent implements OnInit {
+export class ExpensesReportComponent implements OnInit {
 
-  salesIndividual : SalesIndividual[] = [];
-  salesCategories : SalesByCategory[] = [];
-  totalSales : number = 0;
+  expensesIndividual : ExpensesIndividual[] = [];
+  expensesCategories : ExpensesByCategory[] = [];
+  netCosts : number = 0;
   dateFilterEvent:any;
-  constructor(private _service : SalesReportService,
+  constructor(private _service : ExpensesReportService,
               private _errorHandler : ErrorHandlerService,
               private _loader : ProgressDialogService,
               private _router : Router) { }
@@ -36,20 +36,20 @@ export class SalesReportComponent implements OnInit {
       let to   = moment(data.to).format('YYYY-MM-D');
       let type = data.type;
       if(type === 'byIndividual'){
-        this.salesCategories = [];
+        this.expensesCategories = [];
         this._service.getDataIndividual(from,to,type)
         .subscribe( (response) => {
-         this.salesIndividual = response.data; 
-         this.totalSales = response.totalSales;
+         this.expensesIndividual = response.data; 
+         this.netCosts = response.netCosts;
          this._loader.closeSpinner();
         },(err) => { this._errorHandler.errorHandler(err); this._loader.closeSpinner(); });
       }
       else{
-        this.salesIndividual = [];
+        this.expensesIndividual = [];
         this._service.getDataCategory(from,to,type)
         .subscribe( (response) => {
-         this.salesCategories = response.data; 
-         this.totalSales = response.totalSales;
+         this.expensesCategories = response.data; 
+         this.netCosts = response.netCosts;
          this._loader.closeSpinner();
         },(err) => { this._errorHandler.errorHandler(err); this._loader.closeSpinner(); });
       }
@@ -60,11 +60,12 @@ export class SalesReportComponent implements OnInit {
     let from = moment(this.dateFilterEvent.from).format('YYYY-MM-D');
     let to   = moment(this.dateFilterEvent.to).format('YYYY-MM-D'); 
     if(index == 1){
-      this._router.navigateByUrl(`/main/sales/report/print/${from}/${to}/1`);
+      this._router.navigateByUrl(`/main/expenses/report/print/${from}/${to}/1`);
     }
     else{
-      this._router.navigateByUrl(`/main/sales/report/print/${from}/${to}/2`);
+      this._router.navigateByUrl(`/main/expenses/report/print/${from}/${to}/2`);
     }
   }
+
 
 }